+++ entity/calculation.go
@@ -1,6 +1,6 @@
 package entity
 
 type Calculation struct {
-	Expr   string `json:"expr"`
-	Result int64  `json:"result"`
+	Expr   string  `json:"expr"`
+	Result float64 `json:"result"`
 }
